Consolidate TestMeta into test_server.py and remove test_server_interactions.py
